[PORT FROM MAIN]vold: Try to lazy unmount after doUnmount failed
BZ: 102767

Perform a MNT_DETACH lazy unmount after doUnmount failed,
in case that files are been deleting, and the user process
cannot be found and killed by killProcessesWithOpenFiles.

It may not unmount the device right away,
but it will make it unaccessable, and unmount it as soon as it's not
busy.

Change-Id: I4b1f5bd5766f57662c5bf20eded94efd5f3ced1b
Orig-Change-Id: I1575c70f6866919dc3005a89fe54d3393f4a6f79
Category: aosp improvement
Domain: Storage Removable_Storage
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: shuwangx <shux.wang@intel.com>