vold: Try to lazy unmount after doUnmount failed
BZ: 102767

Perform a MNT_DETACH lazy unmount after doUnmount failed,
in case that files are been deleting, and the user process
cannot be found and killed by killProcessesWithOpenFiles.

It may not unmount the device right away,
but it will make it unaccessable, and unmount it as soon as it's not
busy.

Patch has been verified under stress test, perform unmount and mount
while deleting or openning a big directory in a short time.

Change-Id: I1575c70f6866919dc3005a89fe54d3393f4a6f79
Signed-off-by: shuwangx <shux.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/110955
Tested-by: Sun, Jianhua <jianhua.sun@intel.com>
Reviewed-by: Li, Gang G <gang.g.li@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>