[PORT FROM MAIN]vold: Try to lazy unmount after doUnmount failed
BZ: 102767

Perform a MNT_DETACH lazy unmount after doUnmount failed,
in case that files are been deleting, and the user process
cannot be found and killed by killProcessesWithOpenFiles.

It may not unmount the device right away,
but it will make it unaccessable, and unmount it as soon as it's not
busy.

Patch has been verified under stress test, perform unmount and mount
while deleting or openning a big directory in a short time.

Change-Id: I4b1f5bd5766f57662c5bf20eded94efd5f3ced1b
Orig-Change-Id: I1575c70f6866919dc3005a89fe54d3393f4a6f79
Signed-off-by: shuwangx <shux.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/123353
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Huang, Bo B <bo.b.huang@intel.com>
Tested-by: Sun, Jianhua <jianhua.sun@intel.com>
Reviewed-by: Dong, Chuanxiao <chuanxiao.dong@intel.com>