vold:  Resolve Klockwork issue with JB->KK compatibility
BZ: 160701

Klocwork reported an index range issue when moving 2 arrays in a
single call to memmove.  Create a structure to overlay the master key
and salt and then move each of them to the correct location in the
'crypt_mnt_ftr' strucutre.  This is needed when upgrading from
(JB) 1.0 version of the crypto footer to the (KK) 2.0 version.

Change-Id: Iae32dd5121ad2500e6b27ad452d4c7e89c259a9a
Category: aosp improvement
Domain: AOSP-Others
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Norvin Shuster <norvin.shuster@intel.com>